rename form->input, refactor Submission to use getProject
