+++ src/components/ProjectForm.js
@@ -2,13 +2,6 @@ import React from 'react'
 import styled from 'styled-components'
 import { maxWidthMediaQueries } from './Common'
 
-const Container = styled.div`
-  padding: 8vh 16vw;
-  ${maxWidthMediaQueries('mobile')} {
-    padding: 3vh 6vw;
-  }
-`
-
-export default ({ name, devpostLink, videoLink, submitCallback, description }) => {
+export default ({ name, devpostLink, videoLink, description, submitCallback, onChange }) => {
   return <form></form>
 }
