+++ src/containers/Quicklinks.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
-import { Button } from '../components/Form'
+import { Button } from '../components/Input'
 import { db } from '../utility/firebase'
 import { DB_COLLECTION, DB_HACKATHON } from '../utility/Constants'
 import Quicklinks from '../components/Quicklinks'
