+++ src/pages/Submission/Edit.js
@@ -1,8 +1,15 @@
 import React, { useEffect, useState } from 'react'
 import { getLivesiteDoc } from '../../utility/firebase'
+import { Button, TextInput, TextArea, Checkbox, Select, Dropdown } from '../../components/Input'
+import Form from '../../components/ProjectForm'
 
 export default () => {
   const [isSubmissionsOpen, setIsSubmissionsOpen] = useState(false)
+
+  // TODO: use firebasse
+  const [submission, setSubmission] = useState({
+    name: 'Test Submission',
+  })
 
   useEffect(() => {
     const unsubscribe = getLivesiteDoc(livesiteDoc =>
@@ -11,8 +18,12 @@ export default () => {
     return unsubscribe
   }, [setIsSubmissionsOpen])
 
+  const submit = e => {
+    console.log(e)
+  }
+
   return (
     // TODO: Implement this page
-    isSubmissionsOpen ? <h1>TODO: Edit submission form</h1> : <h1>Edit - You cannot submit yet</h1>
+    isSubmissionsOpen ? <Form name={submission.name} /> : <h1>Edit - You cannot submit yet</h1>
   )
 }
