+++ src/pages/Submission/index.js
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react'
-import { getLivesiteDoc, applicantsRef } from '../../utility/firebase'
+import { getLivesiteDoc, applicantsRef, getProject } from '../../utility/firebase'
 import Submission from '../../components/Submission'
-import { getYoutubeThumbnail } from '../../utility/utilities'
 
 // TODO: Get from firebase auth
 const USER_ID = 'aIwA36q0kOw7rDDlCkB2'
@@ -11,34 +10,9 @@ export default () => {
   const [project, setProject] = useState()
   const [feedback, setFeedback] = useState([])
 
-  useEffect(() => {
-    // https://stackoverflow.com/questions/17978883/what-is-the-purpose-of-a-semicolon-before-an-iife
-    ;(async () => {
-      const application = await applicantsRef.doc(USER_ID).get()
-      const team = await application.data().team.get()
-      team
-        .data()
-        .project.get()
-        .then(doc => {
-          const projectData = doc.data()
-          projectData.imgUrl = getYoutubeThumbnail(projectData.youtubeUrl)
-          projectData.href = projectData.devpostUrl
-          setProject(projectData)
-        })
-      team
-        .data()
-        .project.collection('Grades')
-        .orderBy('notes')
-        .get()
-        .then(doc => {
-          const feedback = doc.docs.map(doc => {
-            const docData = doc.data()
-            return docData.notes
-          })
-          setFeedback(feedback)
-        })
-    })()
-  }, [setProject])
+  useEffect(async () => {
+    await getProject(USER_ID, setProject, setFeedback)
+  }, [setProject, setFeedback])
 
   useEffect(() => {
     const unsubscribe = getLivesiteDoc(livesiteDoc =>
